refactor: moved pallet-xdns migration to a separate file
